Fixed issue #1 Parameters sent from client are received empty on the WCF side
The groovy script that builds the Order POJO (wrapper of the Order) object on the WCF service wasn't using the correct methods of the ObjectFactory to set the adequate namespace-aware parameter values. Because of that, the generated SOAP request was sending the values as generic parameters not associated with the object's attributes. The mule flow was updated to build the Order object using the appropiate methods of the ObjectFactory, and the WCF service class was updated to send a message to the console showing the actual values of the received parameters.